tests/update-test-output.py: Adapt to some broken test output
Sometimes, the output of runtestreaddwarf or runtestannotate are
broken due the fact that they execute test units in parallel and each
unit might emit output that watch on each other toes.

This fixes tests/update-test-output.py to take that into account.

As this is a helper tool used to update updates, it won't have any
impact on libabigail's output.

	* tests/update-test-output.py (process): Don't expect the start
	pattern of the main diff hunk to begin at the end of a line
	because that can be broken for runtestreaddwarf and co.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>