test-abidiff-exit: Fix the command line passed to abidiff
I noticed that the command line passed to abidiff in the
test-abidiff-exit.cc is getting bigger and bigger as the test runs.
This is because in the main loop that calls abidiff on each test, we
forget to reset the "headers directories" part of the command line
passed to abidiff.

Fixed thus.

	* tests/test-abidiff-exit.cc (main): Reset the headers directories
	passed to abidiff at each iteration.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>