dwarf-reader: Don't compute canonical type while propagating one
When propagate_canonical_type sees that the source type doesn't have a
canonical type, it tries to compute one for it.

But then suppose we are into computing the canonical type of a type T
to begin with.  At some point we want to try to propagate the
canonical type of a sub-type S of T.  Suppose S recursively references
T.

When propagate_canonical_type is called from
maybe_propagate_canonical_type on S, it's going to try to compute a
canonical type for T (indirectly), as T doesn't yet have a canonical
type.

We then end up in a recursive calling loop that overflows the
execution stack.

	* src/abg-dwarf-reader.cc (propagate_canonical_type): If the
	source (right-hand-side) type has no canonical type, do not
	compute it and thus, do not try to propagate its canonical type to
	the destination (left-hand-side) type.
	* tests/data/test-annotate/test15-pr18892.so.abi: Adjust.
	* tests/data/test-read-dwarf/test15-pr18892.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>