Update the abidw document file
When harmonizing error reporting between abidw and abidiff, I didn't
update the documentation for abidw. This patch adds a return value
section to abidw's man page explaining the error codes.

At the same time, I noticed that the section about Alternate debug
info was out of date and so brought that up to date with the current
versions of the standard.

Signed-off-by: Ben Woodard <woodard@redhat.com>